+++ skforecast/ForecasterAutoregMultiOutput.py
@@ -73,6 +73,10 @@ class ForecasterAutoregMultiOutput():
     max_lag : int
         Maximum value of lag included in lags.
         
+    window_size: int
+        Size of the window needed to create the predictors. It is equal to
+        `max_lag`.
+        
     last_window : 1D np.ndarray
         Last time window the forecaster has seen when trained. It stores the
         values needed to calculate the lags used to predict the next `step`
@@ -86,7 +90,9 @@ class ForecasterAutoregMultiOutput():
             
     exog_shape : tuple
         Shape of exog used in training.
-        
+    
+    fitted: Bool
+        Tag to identify if the estimator is fitted.
         
     Notes
     -----
@@ -106,6 +112,7 @@ class ForecasterAutoregMultiOutput():
         self.included_exog = False
         self.exog_type     = None
         self.exog_shape    = None
+        self.fitted        = False
 
         
         if not isinstance(steps, int) or steps < 1:
@@ -131,7 +138,8 @@ class ForecasterAutoregMultiOutput():
                 f"Got {type(lags)}"
             )
             
-        self.max_lag  = max(self.lags)
+        self.max_lag = max(self.lags)
+        self.window_size = self.max_lag
                 
         
     def __repr__(self) -> str:
@@ -148,6 +156,8 @@ class ForecasterAutoregMultiOutput():
                 + "Steps: " + str(self.steps) \
                 + "\n" \
                 + "Lags: " + str(self.lags) \
+                + "\n" \
+                + "Window size: " + str(self.window_size) \
                 + "\n" \
                 + "Exogenous variable: " + str(self.included_exog) + ', ' + str(self.exog_type) \
                 + "\n" \
@@ -366,6 +376,7 @@ class ForecasterAutoregMultiOutput():
                                             y_train = y_train
                                          ) 
             self.regressors_[step].fit(X_train_step, y_train_step)
+        self.fitted = True
 
         # The last time window of training data is stored so that lags needed as
         # predictors in the first iteration of `predict()` can be calculated.
@@ -408,6 +419,11 @@ class ForecasterAutoregMultiOutput():
             Values predicted.
 
         '''
+
+        if not self.fitted:
+            raise Exception(
+                'This Forecaster instance is not fitted yet. Call `fit` with appropriate arguments before using this it.'
+            )
 
         if exog is None and self.included_exog:
             raise Exception(