+++ skforecast/model_selection.py
@@ -20,11 +20,10 @@ from sklearn.model_selection import ParameterGrid
 
 from .ForecasterAutoreg import ForecasterAutoreg
 from .ForecasterAutoregCustom import ForecasterAutoregCustom
-from .ForecasterCustom import ForecasterCustom
 from .ForecasterAutoregMultiOutput import ForecasterAutoregMultiOutput
 
 logging.basicConfig(
-    format = '%(asctime)-5s %(name)-10s %(levelname)-5s %(message)s', 
+    format = '%(name)-10s %(levelname)-5s %(message)s', 
     level  = logging.INFO,
 )
 
@@ -147,16 +146,14 @@ def cv_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
                   allow_incomplete_fold: bool=True, verbose: bool=True
                  ) -> Tuple[np.array, np.array]:
     '''
-    Cross-validation of `ForecasterAutoreg`, `ForecasterCustom`, `ForecasterAutoregCustom`
+    Cross-validation of `ForecasterAutoreg`, `ForecasterAutoregCustom`
     or `ForecasterAutoregMultiOutput` object. The order of data is maintained
     and the training set increases in each iteration.
     
     Parameters
     ----------
-    forecaster : ForecasterAutoreg, ForecasterCustom, ForecasterAutoregCustom,
-                 ForecasterAutoregMultiOutput
-        `ForecasterAutoreg`, `ForecasterCustom` `ForecasterAutoregCustom` or
-        `ForecasterAutoregMultiOutput` object.
+    forecaster : ForecasterAutoreg, ForecasterAutoregCustom, ForecasterAutoregMultiOutput
+        `ForecasterAutoreg`, `ForecasterAutoregCustom` or `ForecasterAutoregMultiOutput` object.
         
     y : 1D np.ndarray, pd.Series
         Training time series values. 
@@ -202,7 +199,12 @@ def cv_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
     if initial_train_size > len(y):
         raise Exception(
             '`initial_train_size` must be smaller than lenght of `y`.'
-            ' Try to reduce `initial_train_size` or `steps`.'
+        )
+        
+    if initial_train_size is not None and initial_train_size < forecaster.window_size:
+        raise Exception(
+            f"`initial_train_size` must be greater than "
+            f"forecaster's window_size ({forecaster.window_size})."
         )
 
     if metric not in ['mean_squared_error', 'mean_absolute_error',
@@ -262,30 +264,37 @@ def cv_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
 
 
 def backtesting_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
-                           initial_train_size: int, steps: int, metric: str,
+                           steps: int, metric: str, initial_train_size: None,
                            exog: Union[np.ndarray, pd.Series, pd.DataFrame]=None,
                            verbose: bool=False) -> Tuple[np.array, np.array]:
     '''
-    Backtesting (validation) of `ForecasterAutoreg`, `ForecasterCustom`,
-    `ForecasterAutoregCustom` or `ForecasterAutoregMultiOutput` object.
+    Backtesting (validation) of `ForecasterAutoreg`, `ForecasterAutoregCustom` or
+    `ForecasterAutoregMultiOutput` object.
     The model is trained only once using the `initial_train_size` first observations.
-    In each iteration, a number of `steps` predictions are evaluated.
-    
+    In each iteration, a number of `steps` predictions are evaluated.   
     This evaluation is much faster than `cv_forecaster()` since the model is
     trained only once.
+
+    If `forecaster` is already trained and `initial_train_size` is `None`,
+    no initial train is done and all data is used to evaluate the model.
+    However, the first `len(forecaster.last_window)` observations are needed
+    to create the initial predictors, therefore, no predictions are
+    calculated for them.
     
     Parameters
     ----------
-    forecaster : ForecasterAutoreg, ForecasterCustom, ForecasterAutoregCustom,
-                 ForecasterAutoregMultiOutput
-        `ForecasterAutoreg`, `ForecasterCustom` `ForecasterAutoregCustom` or
-        `ForecasterAutoregMultiOutput` object.
+    forecaster : ForecasterAutoreg, ForecasterAutoregCustom, ForecasterAutoregMultiOutput
+        `ForecasterAutoreg`, `ForecasterAutoregCustom` or `ForecasterAutoregMultiOutput` object.
         
     y : 1D np.ndarray, pd.Series
         Training time series values. 
     
-    initial_train_size: int 
-        Number of samples in the initial train split.
+    initial_train_size: int, default `None`
+        Number of samples in the initial train split. If `None` and `forecaster`
+        is already trained, no initial train is done and all data is used to
+        evaluate the model. However, the first `len(forecaster.last_window)`
+        observations are needed to create the initial predictors. Therefore,
+        no predictions are calculated for them.
         
     steps : int
         Number of steps to predict.
@@ -318,10 +327,15 @@ def backtesting_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
         forecaster._check_exog(exog=exog)
         exog = forecaster._preproces_exog(exog=exog)
 
-    if initial_train_size > len(y):
+    if initial_train_size is not None and initial_train_size > len(y):
         raise Exception(
-            '`initial_train_size` must be smaller than lenght of `y`.'
-            ' Try to reduce `initial_train_size` or `steps`.'
+            'If used, `initial_train_size` must be smaller than lenght of `y`.'
+        )
+        
+    if initial_train_size is not None and initial_train_size < forecaster.window_size:
+        raise Exception(
+            f"`initial_train_size` must be greater than "
+            f"forecaster's window_size ({forecaster.window_size})."
         )
 
     if metric not in ['mean_squared_error', 'mean_absolute_error',
@@ -330,8 +344,18 @@ def backtesting_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
             f"Allowed metrics are: 'mean_squared_error', 'mean_absolute_error' and "
             f"'mean_absolute_percentage_error'. Got {metric}."
         )
-    
-    backtest_predictions = []
+
+    if initial_train_size is None and not forecaster.fitted:
+        raise Exception(
+            '`forecaster` must be already trained if no `initial_train_size` is provided.'
+        )
+
+    if initial_train_size is None and forecaster.fitted:
+        logging.warning(
+            f'Altough no initial train is done, the first '
+            f'{len(forecaster.last_window)} observations are needed to create '
+            f'the initial predictors. Therefore, no predictions are calculated for them.'
+        )
     
     metrics = {
         'mean_squared_error': mean_squared_error,
@@ -340,16 +364,21 @@ def backtesting_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
     }
     
     metric = metrics[metric]
-        
-    if exog is None:
-        forecaster.fit(y=y[:initial_train_size])      
+    backtest_predictions = []
+
+    if initial_train_size is not None:
+        if exog is None:
+            forecaster.fit(y=y[:initial_train_size])      
+        else:
+            forecaster.fit(y=y[:initial_train_size], exog=exog[:initial_train_size])
+        window_size = forecaster.window_size
     else:
-        forecaster.fit(y=y[:initial_train_size], exog=exog[:initial_train_size])     
-    
+        # Although not used for training, first observations are needed to create the initial predictors
+        window_size = forecaster.window_size
+        initial_train_size = window_size
     
     folds     = (len(y) - initial_train_size) // steps + 1
     remainder = (len(y) - initial_train_size) % steps
-    window_size = len(forecaster.last_window)
 
     if isinstance(forecaster, ForecasterAutoregMultiOutput) and remainder != 0:
         # In ForecasterAutoregMultiOutput predictions are not iterative,
@@ -361,7 +390,7 @@ def backtesting_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
         remainder=0
     
     if verbose:
-        print(f"Number of observations used for training: {initial_train_size}")
+        print(f"Number of observations used for training or as initial window: {initial_train_size}")
         print(f"Number of observations used for testing: {len(y) - initial_train_size}")
         print(f"    Number of folds: {folds - 1 * (remainder == 0)}")
         print(f"    Number of steps per fold: {steps}")
@@ -371,32 +400,34 @@ def backtesting_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
     for i in range(folds):
         last_window_end   = initial_train_size + i * steps
         last_window_start = (initial_train_size + i * steps) - window_size 
-        last_window       = y[last_window_start:last_window_end]
+        last_window_y     = y[last_window_start:last_window_end]
+        if exog is not None:
+            next_window_exog    = exog[last_window_end:last_window_end + steps]
                 
         if i < folds - 1:
             if exog is None:
                 pred = forecaster.predict(
                             steps       = steps,
-                            last_window = last_window
+                            last_window = last_window_y
                         )
             else:
                 pred = forecaster.predict(
                             steps       = steps,
-                            last_window = last_window,
-                            exog        = exog[last_window_end:last_window_end + steps]
+                            last_window = last_window_y,
+                            exog        = next_window_exog
                         )
         elif remainder != 0:
             steps = remainder 
             if exog is None:
                 pred = forecaster.predict(
                             steps       = steps,
-                            last_window = last_window
+                            last_window = last_window_y
                         )
             else:
                 pred = forecaster.predict(
                             steps       = steps,
-                            last_window = last_window,
-                            exog        = exog[last_window_end:last_window_end + steps]
+                            last_window = last_window_y,
+                            exog        = next_window_exog
                         )
         else:
             continue
@@ -426,10 +457,8 @@ def grid_search_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
     
     Parameters
     ----------
-    forecaster : ForecasterAutoreg, ForecasterCustom, ForecasterAutoregCustom,
-                 ForecasterAutoregMultiOutput
-        `ForecasterAutoreg`, `ForecasterCustom` `ForecasterAutoregCustom` or
-        `ForecasterAutoregMultiOutput` object.
+    forecaster : ForecasterAutoreg, ForecasterAutoregCustom, ForecasterAutoregMultiOutput
+        `ForecasterAutoreg`, `ForecasterAutoregCustom` or `ForecasterAutoregMultiOutput` object.
         
     y : 1D np.ndarray, pd.Series
         Training time series values. 
@@ -486,10 +515,10 @@ def grid_search_forecaster(forecaster, y: Union[np.ndarray, pd.Series],
         forecaster._check_exog(exog=exog)
         exog = forecaster._preproces_exog(exog=exog)
     
-    if isinstance(forecaster, (ForecasterCustom, ForecasterAutoregCustom)):
+    if isinstance(forecaster, ForecasterAutoregCustom):
         if lags_grid is not None:
             logging.warning(
-                '`lags_grid` ignored if forecaster is an instance of `ForecasterCustom` or `ForecasterAutoregCustom`.'
+                '`lags_grid` ignored if forecaster is an instance of `ForecasterAutoregCustom`.'
             )
         lags_grid = ['custom predictors']
         
@@ -577,19 +606,18 @@ def backtesting_forecaster_intervals(forecaster, y: Union[np.ndarray, pd.Series]
                            in_sample_residuals: bool=True,
                            verbose: bool=False) -> Tuple[np.array, np.array]:
     '''
-    Backtesting (validation) of `ForecasterAutoreg`, `ForecasterCustom` or
-    `ForecasterAutoregCustom` object. The model is trained only once using the
-    `initial_train_size` first observations. In each iteration, a number of
-    `steps` predictions are evaluated. Both, predictions and intervals, are
-    calculated.
+    Backtesting (validation) of `ForecasterAutoreg`, or `ForecasterAutoregCustom` object.
+    The model is trained only once using the `initial_train_size` first observations. In 
+    each iteration, a number of `steps` predictions are evaluated. Both, predictions and
+    intervals, are calculated.
     
     This evaluation is much faster than `cv_forecaster()` since the model is
     trained only once.
     
     Parameters
     ----------
-    forecaster : ForecasterAutoreg, ForecasterCustom, ForecasterAutoregCustom
-        `ForecasterAutoreg`, `ForecasterCustom` or `ForecasterAutoregCustom` object.
+    forecaster : ForecasterAutoreg, ForecasterAutoregCustom
+        `ForecasterAutoreg` or `ForecasterAutoregCustom` object.
         
     y : 1D np.ndarray, pd.Series
         Training time series values.