+++ tests/test_skforecast.py
@@ -1,3 +1,6 @@
+import sys
+sys.path.insert(1, '/home/ximo/Documents/GitHub/skforecast')
+
 from skforecast import __version__
 
 def test_version():